refactor: gathering 관련 DTO 클래스 core.domain.gathering.dto 패키지에서 보관하도록 변경
