+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/Gathering.java
@@ -1,5 +1,6 @@
 package com.triplem.momoim.core.domain.gathering;
 
+import com.triplem.momoim.core.domain.gathering.dto.ModifyGathering;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringLocation;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringStatus;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringType;
