+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/gathering/dto/ModifyGathering.java
@@ -1,24 +1,19 @@
-package com.triplem.momoim.core.domain.gathering;
+package com.triplem.momoim.core.domain.gathering.dto;
 
 import com.triplem.momoim.core.domain.gathering.enums.GatheringLocation;
 import com.triplem.momoim.core.domain.gathering.enums.GatheringStatus;
-import com.triplem.momoim.core.domain.gathering.enums.GatheringType;
+import com.triplem.momoim.core.domain.gathering.enums.GatheringSubCategory;
 import java.time.LocalDateTime;
 import java.util.List;
-import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-@AllArgsConstructor
-public class GatheringDetail {
-    private Long id;
-    private Long managerId;
-    private String managerName;
-    private String managerProfileImage;
-    private String category;
-    private String subCategory;
+@Builder
+public class ModifyGathering {
+    private Long gatheringId;
+    private GatheringSubCategory subCategory;
     private String name;
-    private GatheringType gatheringType;
     private GatheringStatus status;
     private String image;
     private String description;
@@ -26,9 +21,6 @@ public class GatheringDetail {
     private List<String> tags;
     private GatheringLocation location;
     private int capacity;
-    private int participantCount;
     private Boolean isPeriodic;
     private LocalDateTime nextGatheringAt;
-    private Boolean isJoined;
-    private Boolean isManager;
 }
