+++ momoim-external-api/src/main/java/com/triplem/momoim/api/review/service/ReviewService.java
@@ -1,7 +1,7 @@
 package com.triplem.momoim.api.review.service;
 
 import com.triplem.momoim.core.common.PaginationInformation;
-import com.triplem.momoim.core.domain.gathering.GatheringPreview;
+import com.triplem.momoim.core.domain.gathering.dto.GatheringPreview;
 import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
 import com.triplem.momoim.core.domain.review.ModifyReview;
 import com.triplem.momoim.core.domain.review.MyReview;
